Add Event Aggregation
Added event publishish in loginviewmodel when user successfully logged on by creating an instance of logonevent class and in shellviewmodel suscribing to the event published and activating the salesviewmodel resulting in activation of salesview. Also removed loginviewmodel private property in shellviewmodel and activated a new instance of loginviewmodel by getting the instance from simple container. Resized shellview width and height and content control margins.