Add API Product Endpoint
Added API endpoint for getting product data from database, added store procedure for it, added product access and models in API library along with UI library API helper for product and model for product. Also added the functionality to add items to the cart and added the cartitemmodel in UI library, get the subtotal and update quantities.